+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_actions.py
@@ -5,7 +5,6 @@
 
 import json
 import re
-from six.moves.urllib.request import urlopen  # pylint: disable=import-error
 
 from knack.util import CLIError
 
@@ -72,14 +71,19 @@ def load_images_thru_services(cli_ctx, publisher, offer, sku, location):
 
 
 def load_images_from_aliases_doc(cli_ctx, publisher=None, offer=None, sku=None):
+    import requests
     from azure.cli.core.cloud import CloudEndpointNotSetException
+    from azure.cli.core.util import should_disable_connection_verify
     try:
         target_url = cli_ctx.cloud.endpoints.vm_image_alias_doc
     except CloudEndpointNotSetException:
         raise CLIError("'endpoint_vm_image_alias_doc' isn't configured. Please invoke 'az cloud update' to configure "
                        "it or use '--all' to retrieve images from server")
-    txt = urlopen(target_url).read()
-    dic = json.loads(txt.decode())
+    # under hack mode(say through proxies with unsigned cert), opt out the cert verification
+    response = requests.get(target_url, verify=(not should_disable_connection_verify()))
+    if response.status_code != 200:
+        raise CLIError("Failed to retrieve image alias doc '{}'. Error: '{}'".format(target_url, response))
+    dic = json.loads(response.content.decode())
     try:
         all_images = []
         result = (dic['outputs']['aliases']['value'])