+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.1.25
+++++++
+* fix broken webapp log tail/download
+* relieve the 'kind' check on webapp/functionapp
 
 0.1.24
 ++++++